Assignment_2/User/requirements.txt
django
djangorestframework
requests
pandas
